18.12W

+++ WebDiet.Server/Services/UserCustomDishService.cs
@@ -23,20 +23,32 @@ namespace WebDiet.Server.Services
             _mapper = mapper;
         }
 
+        public void Update(int id, UserCustomDishDto updatedDish, int userId)
+        {
+            // Pobranie istniejącego dania użytkownika wraz z jego składnikami
+            var userCustomDish = _context.UserCustomDishes
+                .Include(ucd => ucd.CustomIngredients)
+                .FirstOrDefault(ucd => ucd.Id == id && ucd.UserId == userId);
 
+            if (userCustomDish == null)
+            {
+                throw new NotFoundException("User custom dish not found");
+            }
+
+            // Aktualizacja podstawowych właściwości
             userCustomDish.Name = updatedDish.Name ?? userCustomDish.Name;
             userCustomDish.BaseDishId = updatedDish.BaseDishId ?? userCustomDish.BaseDishId;
 
-
+            // Reset wartości odżywczych
             userCustomDish.Protein = 0;
             userCustomDish.Kcal = 0;
             userCustomDish.Fat = 0;
             userCustomDish.Carbo = 0;
 
-
+            // Usunięcie istniejących składników
             _context.UserDishIngredients.RemoveRange(userCustomDish.CustomIngredients);
 
-
+            // Dodanie nowych składników (jeśli są podane)
             if (updatedDish.CustomIngredients != null && updatedDish.CustomIngredients.Any())
             {
                 foreach (var ingredient in updatedDish.CustomIngredients)
@@ -56,13 +68,13 @@ namespace WebDiet.Server.Services
                         Quantity = ingredient.Quantity
                     };
 
-
+                    // Aktualizacja wartości odżywczych
                     userCustomDish.Kcal += (ingredient.Quantity * ingredientEntity.KCal / 100);
                     userCustomDish.Protein += (ingredient.Quantity * ingredientEntity.Protein / 100);
                     userCustomDish.Carbo += (ingredient.Quantity * ingredientEntity.Carbo / 100);
                     userCustomDish.Fat += (ingredient.Quantity * ingredientEntity.Fat / 100);
 
-
+                    // Dodanie nowego składnika do kontekstu
                     _context.UserDishIngredients.Add(userDishIngredient);
                 }
             }
@@ -70,7 +82,7 @@ namespace WebDiet.Server.Services
             // Zapis zmian
             _context.SaveChanges();
         }
-        public UserCustomDishDto GetById(int id)
+        public DishDto GetById(int id)
         {
             var userCustomDish = _context.UserCustomDishes
                 .Include(ci => ci.CustomIngredients)
